Updated based on review * Renamed the Adaptors Data Access View package from constellations to constelletion * Removed and excluded some more dependencies. Testing via the unit test seems to work OK.
