+++ AdaptorsDataAccessPlugins/manifest.mf
@@ -1,0 +1,5 @@
+Manifest-Version: 1.0
+OpenIDE-Module: au.gov.asd.tac.constellation.dataaccess.adaptors
+OpenIDE-Module-Localizing-Bundle: au/gov/asd/tac/constellation/dataaccess/adaptors/Bundle.properties
+OpenIDE-Module-Specification-Version: 1.0
+
