+++ AdaptorsDataAccessPlugins/nbproject/project.xml
@@ -3,7 +3,7 @@
     <type>org.netbeans.modules.apisupport.project</type>
     <configuration>
         <data xmlns="http://www.netbeans.org/ns/nb-module-project/3">
-            <code-name-base>au.gov.asd.tac.constellations.dataaccess.adaptors</code-name-base>
+            <code-name-base>au.gov.asd.tac.constellation.dataaccess.adaptors</code-name-base>
             <suite-component/>
             <module-dependencies>
                 <dependency>
