+++ AdaptorsDataAccessPlugins/src/au/gov/asd/tac/constellation/dataaccess/adaptors/Bundle.properties
@@ -1,0 +1,4 @@
+OpenIDE-Module-Name=Adaptors Data Access Plugins
+OpenIDE-Module-Display-Category=Adaptors
+OpenIDE-Module-Long-Description=Adaptors for the Data Access View.
+OpenIDE-Module-Short-Description=Adaptors Data Access
