--- AdaptorsDataAccessPlugins/src/au/gov/asd/tac/constellations/dataaccess/adaptors/Bundle.properties
@@ -1,1 +1,0 @@
-OpenIDE-Module-Name=Adaptors Data Access Plugins
