+++ AdaptorsDataAccessPlugins/test/unit/src/au/gov/asd/tac/constellation/dataaccess/adaptors/au/gov/asd/tac/constellations/dataaccess/adaptors/providers/gaffer/example/resources/exampleGafferResponseTwoHop.json
