+++ nbproject/project.properties
@@ -15,11 +15,9 @@ modules=\
     ${project.au.gov.asd.tac.constellation.views.mapview.adaptors}:\
     ${project.au.gov.asd.tac.constellation.dependencies.adaptors}:\
     ${project.au.gov.asd.tac.constellation.functionality.adaptors}:\
-    ${project.au.gov.asd.tac.constellations.dataaccess.adaptors}
+    ${project.au.gov.asd.tac.constellation.dataaccess.adaptors}
 
 project.au.gov.asd.tac.constellation.dependencies.adaptors=AdaptorsDependencies
 project.au.gov.asd.tac.constellation.functionality.adaptors=AdaptorsFunctionality
 project.au.gov.asd.tac.constellation.views.mapview.adaptors=AdaptorsMapView
-project.au.gov.asd.tac.constellations.dataaccess.adaptors=AdaptorsDataAccessPlugins
-
-app.conf=constellation.conf
+project.au.gov.asd.tac.constellation.dataaccess.adaptors=AdaptorsDataAccessPlugins
